+++ CalligraphySample/src/main/java/uk/co/chrisjenx/calligraphy/sample/MainActivity.java
@@ -2,12 +2,17 @@ package uk.co.chrisjenx.calligraphy.sample;
 
 import android.content.Context;
 import android.os.Bundle;
+import android.os.Handler;
+import android.os.Looper;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentActivity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Toast;
 
+import butterknife.ButterKnife;
+import butterknife.OnClick;
 import uk.co.chrisjenx.calligraphy.CalligraphyConfig;
 import uk.co.chrisjenx.calligraphy.CalligraphyContextWrapper;
 
@@ -19,6 +24,14 @@ public class MainActivity extends FragmentActivity {
         CalligraphyConfig.initDefault("fonts/Roboto-ThinItalic.ttf");
 
         setContentView(R.layout.activity_main);
+
+        new Handler(Looper.getMainLooper()).postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                setTitle("Calligraphy changed");
+                getActionBar().setSubtitle("Added subtitle");
+            }
+        }, 1000);
     }
 
     @Override
@@ -30,9 +43,21 @@ public class MainActivity extends FragmentActivity {
      * A placeholder fragment containing a simple view.
      */
     public static class PlaceholderFragment extends Fragment {
+
         @Override
         public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle state) {
             return inflater.inflate(R.layout.fragment_main, container, false);
+        }
+
+        @Override
+        public void onViewCreated(View view, Bundle savedInstanceState) {
+            super.onViewCreated(view, savedInstanceState);
+            ButterKnife.inject(this, view);
+        }
+
+        @OnClick({R.id.button_default, R.id.button_bold})
+        public void onClickButton() {
+            Toast.makeText(getActivity(), "Custom Typeface toast text", Toast.LENGTH_SHORT).show();
         }
     }
 }