+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 #Changelog
 
 #0.8.0 (WIP)
+- Added ActionBar Title/SubTitle support.
+- Toast support via default style/or TextView theme style.
 
 #0.7.1 (22/04/2014)
-- Fixed Resources not found Exception - PR#31 [@Smuldr](https://github.com/Smuldr)
+- Fixed Resources not found Exception - [@Smuldr](https://github.com/Smuldr)
 
 #0.7.0 (28/01/2014)
 - Added Anti-aliasing support